# openapi.yaml
openapi: 3.0.3
info:
  title: Mercado Uno API
  version: 0.2.0
  description: |
    API REST del marketplace educativo Mercado Uno. PostgreSQL es la base de
    producción. El checkout reserva inventario y crea un pedido y un pago
    pendientes; el proveedor simulado solo está disponible fuera de producción.
  license:
    name: Uso educativo
security: []
servers:
  - url: http://localhost:3000
    description: Stack Docker, con la API publicada por el proxy del frontend
  - url: http://localhost:4000
    description: API directa para desarrollo
...
paths:
...
  /api/checkout:
    post:
      tags: [Orders]
      summary: Reservar inventario y crear un pedido pendiente
      operationId: checkoutCart
      description: No cobra ni marca la orden como pagada. Una misma clave reutiliza el resultado anterior.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Pedido nuevo o reutilizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
  /api/orders:
    get:
      tags: [Orders]
      summary: Listar pedidos propios
      operationId: listOrders
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Pedidos del usuario autenticado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Obtener un pedido propio
      operationId: getOrder
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Pedido, líneas y resumen de pago.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/InvalidIdentifier"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orders/{orderId}/pay:
    post:
      tags: [Orders]
      summary: Ejecutar el proveedor simulado
      operationId: payOrder
      description: Solo está habilitado en desarrollo y tests; es idempotente por estado.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: string
                  enum: [paid, failed]
                  default: paid
      responses:
        "200":
          description: Resultado nuevo o reutilizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          description: El proveedor simulado está deshabilitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancelar un pedido y liberar su inventario una vez
      operationId: cancelOrder
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Cancelación nueva o reutilizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderMutationResponse"
        "400":
          $ref: "#/components/responses/InvalidIdentifier"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidOrigin"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mercado_session
      description: En producción se usa el nombre endurecido __Host-mercado_session.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Compatibilidad temporal para clientes no navegador.
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  responses:
    BadRequest:
      description: Entrada inválida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidIdentifier:
      description: El identificador no es un entero positivo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Sesión ausente, inválida o expirada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidOrigin:
      description: Una mutación con cookie llegó desde un origen no autorizado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: El recurso no existe o no pertenece al usuario.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: El estado actual no permite la transición.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    OrderItem:
      type: object
      required: [id, ProductId, qty, price]
      properties:
        id: { type: integer }
        ProductId: { type: integer }
        qty: { type: integer }
        price: { type: number, format: double }
    PaymentSummary:
      type: object
      nullable: true
      required: [status, provider, amountCents, currency]
      properties:
        status:
          type: string
          enum: [pending, paid, failed, cancelled, refunded]
        provider: { type: string, enum: [simulated] }
        amountCents: { type: integer, minimum: 0 }
        currency: { type: string, example: MXN }
    Order:
      type: object
      required: [id, status, total, items, payment]
      properties:
        id: { type: integer }
        status:
          type: string
          enum: [pending_payment, paid, payment_failed, cancelled]
        total: { type: number, format: double }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        payment:
          $ref: "#/components/schemas/PaymentSummary"
    OrderMutationResponse:
      type: object
      required: [ok, replayed, order]
      properties:
        ok: { type: boolean, enum: [true] }
        orderId:
          type: integer
          description: Alias conservado en la respuesta de checkout.
        replayed: { type: boolean }
        order:
          $ref: "#/components/schemas/Order"